feat: [CVNG-2926]: completing integration of Health Sources component to Verify Step (#4497)
* feat: [CVNG-2941]: fixing linting issues

* feat: [CVNG-2926]: completing integration of Health Sources component to Verify Step

* feat: [CVNG-2926]: fixing testcases

* feat: [CVNG-2926]: fixing issue of adding health source immediately once monitoring source is created

+++ src/modules/85-cv/pages/health-source/HealthSourceTable/HealthSourceTable.tsx
@@ -2,7 +2,7 @@ import React, { useState, useMemo, useCallback } from 'react'
 import cx from 'classnames'
 import { Link, useParams, useHistory } from 'react-router-dom'
 import type { CellProps, Renderer } from 'react-table'
-import { Color, Container, Text, SelectOption } from '@wings-software/uicore'
+import { Color, Container, Text } from '@wings-software/uicore'
 import { useToaster } from '@common/exports'
 import { NoDataCard } from '@common/components/Page/NoDataCard'
 import type { ProjectPathProps } from '@common/interfaces/RouteInterfaces'
@@ -12,23 +12,11 @@ import routes from '@common/RouteDefinitions'
 import { Table } from '@common/components'
 import ContextMenuActions from '@cv/components/ContextMenuActions/ContextMenuActions'
 import { getIconBySourceType } from '@cv/pages/admin/setup/SetupUtils'
+import HealthSources from '@cv/components/PipelineSteps/ContinousVerification/components/HealthSources/HealthSources'
 import HealthSourceDrawerContent from '../HealthSourceDrawer/HealthSourceDrawerContent'
 import type { RowData } from '../HealthSourceDrawer/HealthSourceDrawerContent.types'
+import type { HealthSourceTableInterface } from './HealthSourceTable.types'
 import css from './HealthSourceTable.module.scss'
-
-interface HealthSourceTableInterface {
-  value: Array<RowData>
-  breadCrumbRoute?: {
-    routeTitle: string
-    redirect: () => void
-  }
-  monitoringSourcRef: { monitoredServiceIdentifier: string; monitoredServiceName: string }
-  serviceRef: SelectOption | undefined
-  environmentRef: SelectOption | undefined
-  onSuccess: (value: MonitoredServiceResponse) => void
-  onDelete: (value: MonitoredServiceResponse) => void
-  isEdit?: boolean
-}
 
 export default function HealthSourceTable({
   breadCrumbRoute,
@@ -38,14 +26,16 @@ export default function HealthSourceTable({
   value: tableData,
   onSuccess,
   onDelete,
-  isEdit
+  isEdit,
+  shouldRenderAtVerifyStep,
+  isRunTimeInput
 }: HealthSourceTableInterface): JSX.Element {
   const [modalOpen, setModalOpen] = useState(false)
   const history = useHistory()
   const { showError } = useToaster()
   const params = useParams<ProjectPathProps & { identifier: string }>()
   const { getString } = useStrings()
-  const { routeTitle, redirect } = breadCrumbRoute || {}
+  const { routeTitle } = breadCrumbRoute || {}
   const [rowData, setrowData] = useState<RowData | null>(null)
 
   const { mutate: updateMonitoredService } = useUpdateMonitoredService({
@@ -57,12 +47,14 @@ export default function HealthSourceTable({
     return (
       <>
         <Text
+          className={css.breadCrumbLink}
+          style={{ cursor: 'pointer' }}
           icon={'arrow-left'}
           iconProps={{ color: Color.PRIMARY_7, margin: { right: 'small' } }}
           color={Color.PRIMARY_7}
           onClick={() => {
-            redirect
-              ? redirect()
+            shouldRenderAtVerifyStep
+              ? setModalOpen(false)
               : history.push(
                   routes.toCVMonitoringServices({
                     orgIdentifier: params.orgIdentifier,
@@ -83,37 +75,40 @@ export default function HealthSourceTable({
         </div>
       </>
     )
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [rowData, isEdit])
 
-  const deleteHealthSource = async (selectedRow: RowData) => {
-    try {
-      const payload: MonitoredServiceDTO = {
-        orgIdentifier: params.orgIdentifier,
-        projectIdentifier: params.projectIdentifier,
-        serviceRef: selectedRow.service as string,
-        environmentRef: selectedRow.environment as string,
-        identifier: monitoringSourcRef?.monitoredServiceIdentifier,
-        name: monitoringSourcRef?.monitoredServiceName,
-        description: 'monitoredService',
-        type: 'Application',
-        tags: {},
-        sources: {
-          healthSources: tableData?.filter(healthSource => healthSource.identifier !== selectedRow.identifier)
+  const deleteHealthSource = async (selectedRow: RowData): Promise<void> => {
+    if (onDelete) {
+      try {
+        const payload: MonitoredServiceDTO = {
+          orgIdentifier: params.orgIdentifier,
+          projectIdentifier: params.projectIdentifier,
+          serviceRef: selectedRow.service as string,
+          environmentRef: selectedRow.environment as string,
+          identifier: monitoringSourcRef?.monitoredServiceIdentifier,
+          name: monitoringSourcRef?.monitoredServiceName,
+          description: 'monitoredService',
+          type: 'Application',
+          tags: {},
+          sources: {
+            healthSources: tableData?.filter(healthSource => healthSource.identifier !== selectedRow.identifier)
+          }
         }
+        const deletePayload = await updateMonitoredService(payload)
+        await onDelete(deletePayload?.resource as MonitoredServiceResponse)
+      } catch (error) {
+        showError(error?.message)
       }
-      const deletePayload = await updateMonitoredService(payload)
-      await onDelete(deletePayload?.resource as MonitoredServiceResponse)
-    } catch (error) {
-      showError(error?.message)
     }
   }
 
-  const onSuccessHealthSourceTableWrapper = (data: MonitoredServiceResponse) => {
+  const onSuccessHealthSourceTableWrapper = (data: MonitoredServiceResponse): void => {
     setModalOpen(false)
     onSuccess(data)
   }
 
-  const onCloseHealthSourceTableWrapper = () => {
+  const onCloseHealthSourceTableWrapper = (): void => {
     setModalOpen(false)
     setrowData(null)
   }
@@ -135,71 +130,91 @@ export default function HealthSourceTable({
           onEdit={() => {
             const rowFilteredData =
               tableData?.find((healthSource: RowData) => healthSource.identifier === rowdata.identifier) || null
-            rowFilteredData && setrowData(rowFilteredData)
-            setModalOpen(true)
+            editRow(rowFilteredData)
           }}
         />
       </Container>
     )
   }
 
+  const editRow = useCallback(rowToEdit => {
+    if (rowToEdit) {
+      rowToEdit && setrowData(rowToEdit)
+      setModalOpen(true)
+    } else {
+      showError(getString('cv.healthSource.noDataPresentHealthSource'))
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
+
   const disableAddNewHealthSource = useMemo(() => {
     return !monitoringSourcRef?.monitoredServiceName || !serviceRef?.value || !environmentRef?.value
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [monitoringSourcRef, serviceRef, serviceRef])
 
   return (
     <>
-      <Text className={css.tableTitle}>{getString('connectors.cdng.healthSources.label')}</Text>
-      {tableData.length ? (
-        <Table
-          className={css.tableWrapper}
-          sortable={true}
-          onRowClick={data => {
-            const rowFilteredData =
-              tableData?.find((healthSource: RowData) => healthSource.identifier === data.identifier) || null
-            rowFilteredData && setrowData(rowFilteredData)
-            setModalOpen(true)
-          }}
-          columns={[
-            {
-              Header: getString('name'),
-              accessor: 'name',
-              width: '15%'
-            },
-            {
-              Header: getString('typeLabel'),
-              width: '15%'
-            },
-            {
-              Header: getString('source'),
-              accessor: 'type',
-              width: '15%',
-              Cell: renderTypeWithIcon
-            },
-            {
-              Header: getString('cv.healthSource.table.environmentMapping'),
-              accessor: 'environment',
-              width: '20%'
-            },
-            {
-              Header: getString('cv.healthSource.table.serviceMapping'),
-              accessor: 'service',
-              Cell: renderEditDelete,
-              width: '35%'
-            }
-          ]}
-          data={tableData}
+      {shouldRenderAtVerifyStep ? (
+        <HealthSources
+          healthSources={tableData}
+          editHealthSource={editRow}
+          addHealthSource={() => setModalOpen(true)}
+          isRunTimeInput={isRunTimeInput}
         />
       ) : (
-        <Container className={css.noData}>
-          <NoDataCard icon={'join-table'} message={getString('cv.healthSource.noData')} />
-        </Container>
+        <>
+          <Text className={css.tableTitle}>{getString('connectors.cdng.healthSources.label')}</Text>
+          {tableData.length ? (
+            <Table
+              className={css.tableWrapper}
+              sortable={true}
+              onRowClick={data => {
+                const rowFilteredData =
+                  tableData?.find((healthSource: RowData) => healthSource.identifier === data.identifier) || null
+                editRow(rowFilteredData)
+              }}
+              columns={[
+                {
+                  Header: getString('name'),
+                  accessor: 'name',
+                  width: '15%'
+                },
+                {
+                  Header: getString('typeLabel'),
+                  width: '15%'
+                },
+                {
+                  Header: getString('source'),
+                  accessor: 'type',
+                  width: '15%',
+                  Cell: renderTypeWithIcon
+                },
+                {
+                  Header: getString('cv.healthSource.table.environmentMapping'),
+                  accessor: 'environment',
+                  width: '20%'
+                },
+                {
+                  Header: getString('cv.healthSource.table.serviceMapping'),
+                  accessor: 'service',
+                  Cell: renderEditDelete,
+                  width: '35%'
+                }
+              ]}
+              data={tableData}
+            />
+          ) : (
+            <Container className={css.noData}>
+              <NoDataCard icon={'join-table'} message={getString('cv.healthSource.noData')} />
+            </Container>
+          )}
+          <div className={cx(css.drawerlink, disableAddNewHealthSource && css.disabled)}>
+            <Link to={'#'} onClick={() => setModalOpen(true)}>
+              + {getString('cv.healthSource.addHealthSource')}
+            </Link>
+          </div>
+        </>
       )}
-      <div className={cx(css.drawerlink, disableAddNewHealthSource && css.disabled)}>
-        <Link to={'#'} onClick={() => setModalOpen(true)}>
-          + {getString('cv.healthSource.addHealthSource')}
-        </Link>
-      </div>
       <HealthSourceDrawerContent
         isEdit={!!isEdit && !!rowData}
         onClose={onCloseHealthSourceTableWrapper}
@@ -211,6 +226,7 @@ export default function HealthSourceTable({
         monitoringSourcRef={monitoringSourcRef}
         serviceRef={serviceRef}
         environmentRef={environmentRef}
+        shouldRenderAtVerifyStep={shouldRenderAtVerifyStep}
       />
     </>
   )
